Redeploy contracts from the same deployer (#59)
There have been instances of a user sending ETH to the ETH-flow address
in the wrong network, which has a high chance of making the funds
completely unrecoverable.

To reduce concerns of this kind, the ETH-flow contract has been
redeployed to the same address on all supported chains. There is still a
risk when sending a transaction for a network to another one, but in
many cases the order should just expire and the refunder should
eventually reimburse the user. Note that the deployed contract doesn't
use CREATE2, meaning that there's no guarantee that the contract at the
target address is indeed the ETH-flow contract code.

I also added the standard json representation of the solc input (and
successfully tested it). If Etherscan verification through the
verification script fails (unfortunately, a common occurrence) it should
be possible to give verification another shot through the Etherscan web
interface.

### How to test

Confirm that the barn deployment
(0x04501b9b1d52e67f6862d157e00d13419d2d6e95) and the prod one
(0xba3cb449bd2b4adddbc894d8697f5170800eadec) exist on all networks, that
their contract code is verified matches the ETH flow, and that the
deployment parameters are the expected ones.