Fixes for Node 20 and latest Foundry support (#60)
CI failed on the latest push to master. This has two reasons:

-
\[[failure](https://github.com/cowprotocol/ethflowcontract/actions/runs/13183055660/job/36798434698)\]
The Foundry version is updated automatically and had a breaking change
(see
[here](https://book.getfoundry.sh/guides/v1.0-migration#expect-revert-cheatcode-disabled-on-internal-calls-by-default)).
This is somewhat nice as it surfaces the fact that a test wasn't
actually working as expected (because of a Foundry bug though).

-
\[[failure](https://github.com/cowprotocol/ethflowcontract/actions/runs/13183055657/job/36798434017)\]
The latest version of Node isn't compatible with the pinned package in
our artifact generation procedure. I just pinned the Node version as
well because I hope we're going to remove Hardhat artifact generation
eventually anyway.
(I tried to update Hardhat to v2.17.0 ([the earliest that supports Node
20](https://github.com/NomicFoundation/hardhat/blob/dd19b668e3a68085eea87f96dc05e65ae52f0ce3/packages/hardhat-core/CHANGELOG.md#2170))
but I got another error (`HardhatError: HH502: Couldn't download
compiler version list. Please check your internet connection and try
again`) which would have taken too much time to debug compared to trying
to remove Hardhat in the first place.


### How to test

I released a tag to test artifact generation. The result can be found
[here](https://github.com/cowprotocol/ethflowcontract/tree/test-ci) (the
tag will be deleted once the PR is merged), note that CI is successful.