Use obfuscated key in deployment (#68)
A very simple mechanism to make sure that the deployer key for the ETH
flow isn't used for deploying unrelated contracts.

The ETH-flow contract deployment is notoriously fragile because it
relies on CREATE and not CREATE2.
This means that the deployment address only depends on the deployer
address and its current nonce.
If the same deployer/nonce combination is used for any transaction other
than deploying this contract, there's no chance to ever get the same
contract at the same address on that chain.
What's worse, if users by accident send funds to the wrong ETH-flow
contract address, the funds are likely to be permanently lost.

The approach of this PR is using a different deployment procedure.
Instead of using the private key as a parameter of the script, as
customary in most deployment scripts, it uses an obfuscated version of
the private key.
The idea is that having a non-standard format for the private key
prevents it from being used by accident for other deployments.
This is _not_ intended to be a way to hide the key: the private key is
still trivially derivable from the obfuscated version.

This PR:
- Updates the readme.
- Updates the deployment script.
- Adds a script to obfuscate a key.

There are also two tangential changes that I found helpful and decided
to include as well:
- Update forge-std to
[v1.11.0](https://github.com/foundry-rs/forge-std/releases/tag/v1.11.0):
this gives access to `vm.toString` and `Vm.Wallet`.
- Include `script/` in linting.

### How to test

First, try to obfuscate a private key you control.
Then, use that obfuscated key to test a deployment on some testnet.
Let me know how to update the readme if something isn't clear.